refactor(ai): extract translation handler
Move translation model orchestration into a dedicated handler and update dependency injection and consumers.
