+++ app/src/main/java/me/rerere/rikkahub/di/AppModule.kt
@@ -85,6 +85,7 @@ val appModule = module {
             conversationRepo = get(),
             memoryRepository = get(),
             generationHandler = get(),
+            translationHandler = get(),
             templateTransformer = get(),
             providerManager = get(),
             localTools = get(),
