+++ app/src/main/java/me/rerere/rikkahub/di/DataSourceModule.kt
@@ -18,6 +18,7 @@ import me.rerere.rikkahub.data.ai.AIRequestInterceptor
 import me.rerere.rikkahub.data.ai.RequestLoggingInterceptor
 import me.rerere.rikkahub.data.ai.transformers.AssistantTemplateLoader
 import me.rerere.rikkahub.data.ai.GenerationHandler
+import me.rerere.rikkahub.data.ai.TranslationHandler
 import me.rerere.rikkahub.data.ai.transformers.TemplateTransformer
 import me.rerere.rikkahub.data.api.RikkaHubAPI
 import me.rerere.rikkahub.data.api.SponsorAPI
@@ -163,6 +164,10 @@ val dataSourceModule = module {
             json = get(),
             memoryRepo = get()
         )
+    }
+
+    single {
+        TranslationHandler(providerManager = get())
     }
 
     single<OkHttpClient> {
